+++ lib/soundtrack.js
@@ -4,49 +4,112 @@ var rest = require('restler');
 var async = require('async');
 var slug = require('speakingurl');
 
-var Soundtrack = function(app) {
+var Soundtrack = function( config ) {
   var self = this;
-
-  this.app = app;
-  this.app.rooms = {};
-  this.backupTracks = [];
-  this.timers = {
+  self.config = config;
+  self.rooms = {};
+  self.backupTracks = [];
+  self.timers = {
     scrobble: {}
   };
 
-  this.DEBUG = false;
+  self.DEBUG = false;
 
 };
 
-Soundtrack.prototype.start = function() {
+Soundtrack.prototype.start = function( done ) {
+  if (!done) var done = new Function();
   var self = this;
-  // periodically check for idle sockets
-  setInterval(function() {
-    self.markAndSweep();
-  }, self.app.config.connection.checkInterval );
-};
+  
+  self.sub.on('message', function(channel, msg) {
+    console.log( msg );
+    console.log('ORCH received message on channel', channel , JSON.parse(msg).type );
 
-Soundtrack.prototype.sortPlaylist = function() {
-  var self = this;
-  var app = this.app;
-  app.room.playlist = _.union( [ app.room.playlist[0] ] , app.room.playlist.slice(1).sort(function(a, b) {
-    if (b.score === a.score) {
-      return a.timestamp - b.timestamp;
-    } else {
-      return b.score - a.score;
+    var msg = JSON.parse( msg );
+    console.log( msg.data );
+        
+    switch (msg.type) {
+      case 'skip':
+        self.rooms[ msg.data.room ].nextSong();
+        break;
+      case 'queue':
+        console.log('queue will be:', msg.data );
+        self.trackFromSource( msg.data.source , msg.data.id , function(err, track) {
+          console.log('trackFromSource() callback executing...', err || track._id );
+          if (err || !track) return console.error(err);
+          
+          var queueWasEmpty = false;
+          if (!self.rooms[ msg.data.room ].playlist.length) queueWasEmpty = true;
+          self.rooms[ msg.data.room ].queueTrack(track, msg.data.user, function() {
+            console.log( 'queueTrack() callback executing... ');
+
+            if (queueWasEmpty) {
+              self.rooms[ msg.data.room ].broadcast({
+                type: 'track',
+                data: track,
+                //sources: sources,
+                seekTo: 0.0
+              });
+            }
+          });
+        });
+        break;
     }
-  }) );
-}
+  });
+  
+  // periodically check for idle sockets
+  /* setInterval(function() {
+    self.markAndSweep();
+  }, self.config.connection.checkInterval );*/
+
+  Room.find().exec(function(err, rooms) {
+    async.map( rooms , function(room, complete) {
+      self.sub.get( self.config.database.name + ':rooms:' + room.slug + ':playlist', function(err, playlist) {
+        playlist = JSON.parse(playlist);
+        
+        room.playlist = playlist;
+        
+        if (!playlist || !playlist.length) playlist = [];
+        
+        self.rooms[ room.slug ] = room;
+        self.rooms[ room.slug ].playlist = playlist;
+        self.rooms[ room.slug ].listeners = {};
+        
+        self.rooms[ room.slug ].bind( self );
+        
+        self.sub.subscribe('soundtrack:rooms:' + room.slug );
+        self.sub.subscribe('soundtrack:rooms:' + room.slug + ':*' );
+        
+        function errorHandler(err) {
+          if (err) {
+            return self.rooms[ room.slug ].retryTimer = setTimeout(function() {
+              self.rooms[ room.slug ].startMusic( errorHandler );
+            }, 5000 );
+          }
+          
+          return done();
+        }
+        
+        self.rooms[ room.slug ].startMusic( errorHandler );
+
+      });
+    }, function(err, results) {
+      
+    });
+  });
+  
+};
+
 Soundtrack.prototype.broadcast = function(msg) {
   var self = this;
-  var app = this.app;
+
   switch (msg.type) {
     // special handling for edits
     // this allows us to simply update the in-memory version of a track,
     // rather than querying mongoDB several more times for data.
     case 'edit':
-      for (var roomName in app.rooms) {
-        var room = app.rooms[ roomName ];
+      for (var roomName in self.rooms) {
+        var room = self.rooms[ roomName ];
         for (var i = 0; i < room.playlist.length; i++) {
           if ( room.playlist[ i ]._id.toString() == msg.track._id.toString() ) {
             room.playlist[ i ].title        = msg.track.title;
@@ -59,13 +122,10 @@ Soundtrack.prototype.broadcast = function(msg) {
       }
     break;
   }
+  
+  var json = JSON.stringify( msg );
+  self.pub.publish('soundtrack:messages', json );
 
-  //app.redis.set(app.config.database.name + ':playlist', JSON.stringify( app.room.playlist ) );
-
-  var json = JSON.stringify(msg);
-  for (var id in app.clients) {
-    app.clients[id].write(json);
-  }
 };
 Soundtrack.prototype.whisper = function(id, msg) {
   var self = this;
@@ -277,7 +337,7 @@ Soundtrack.prototype.trackFromSource = function(source, id, sourceCallback) {
       });
     break;
     case 'soundcloud':
-      rest.get('https://api.soundcloud.com/tracks/'+parseInt(id)+'.json?client_id='+app.config.soundcloud.id).on('complete', function(data, response) {
+      rest.get('https://api.soundcloud.com/tracks/'+parseInt(id)+'.json?client_id='+ self.config.soundcloud.id).on('complete', function(data, response) {
         console.log(response);
         if (!data.title) return sourceCallback('No Soundcloud track found in ' , data );
